Move fonts test to separate module
